DEBUG example:  bug in getOutputFiles(). Getting the wrong names of the write functions. Including all the read functions.
call with
o = getOutputFiles("code")
from within the variety_trial_analysis/ directory.

The bug is the missing i in primiiveFuns in the  findReadDataFuns.expression method.
Since this doesn't exactly match it is part of the ... call and that is how we get both the PrimitiveReadDataFuns and also the PrimitiveWriteDataFuns mixed together and we see the error in a call to excel_sheets(path = pp).
And that doesn't have a file argument to match so we get an error in findLiteralValue for a call to getCallParam().